Destop
This project consists of an HTML form designed to collect information related to environmental awareness. The form prompts users to input their name, age, contact information, city name, and address. Additionally, it includes questions about the user's awareness of environmental issues and their involvement in environmental initiatives or organizations.
The form is styled using CSS to provide an engaging and visually appealing interface. It utilizes various input elements such as text fields, radio buttons, and a textarea for feedback. The design aims to encourage user participation and facilitate the collection of valuable insights into environmental awareness and engagement.